Abstract Duel class used by all duel events Duel class contains all duel methods which are used by all duel events
