+++ src/me/nicbo/InvadedLandsEvents/EventsMain.java
@@ -16,7 +16,7 @@ import java.util.logging.Logger;
  * Main class
  *
  * @author Nicbo
- * @author StarZorroww
+ * @author StarZorrow
  * @since 2020-03-12
  */
