+++ src/me/nicbo/InvadedLandsEvents/events/duels/sumo/Sumo.java
@@ -1,9 +1,9 @@
-package me.nicbo.InvadedLandsEvents.events.sumo;
+package me.nicbo.InvadedLandsEvents.events.duels.sumo;
 
 import me.nicbo.InvadedLandsEvents.EventsMain;
-import me.nicbo.InvadedLandsEvents.events.InvadedEvent;
+import me.nicbo.InvadedLandsEvents.events.duels.Duel;
 
-public class Sumo extends InvadedEvent {
+public class Sumo extends Duel {
     Sumo(String name, EventsMain plugin, int teamSize) {
         super(name, "sumo", plugin);
     }
@@ -29,7 +29,7 @@ public class Sumo extends InvadedEvent {
     }
 
     @Override
-    public void over() {
+    public void applyKitandTp() {
 
     }
 }
