+++ src/me/nicbo/InvadedLandsEvents/messages/EventMessage.java
@@ -7,7 +7,7 @@ import org.bukkit.ChatColor;
  * All event messages
  *
  * @author Nicbo
- * @author StarZorroww
+ * @author StarZorrow
  * @author thehydrogen
  * @since 2020-03-12
  */
