+++ src/me/nicbo/InvadedLandsEvents/messages/EventPartyMessage.java
@@ -7,7 +7,7 @@ import org.bukkit.ChatColor;
  * All event party messages
  *
  * @author Nicbo
- * @author StarZorroww
+ * @author StarZorrow
  * @since 2020-03-12
  */
