+++ src/me/nicbo/InvadedLandsEvents/party/EventParty.java
@@ -7,7 +7,7 @@ import org.bukkit.entity.*;
 /**
  * Party class
  *
- * @author StarZorroww
+ * @author StarZorrow
  * @since 2020-03-12
  */
